+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v1.1.10
- - Fixed bug causing user's custom title not to show in-game if it's non-RGB (maybe)
+ - Fixed bug causing user's custom title not to show in game if it's non-RGB (maybe)
  - Added colored console logs
  - Updated internal SDK to match latest version of RL (v2.64)